backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-jwt-extended>=4.6
bcrypt>=4.1
Pillow>=10.0
